# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Relation Graph API",
    "version": "0.1.5-acc.52"
  },
  "servers": [
    {
      "url": "https://{tenant}.continuousc.contc",
      "variables": {
        "tenant": {
          "enum": [
            "tenant-demo",
            "tenant-mdp"
          ],
          "default": "tenant-demo"
        }
      }
    }
  ],
  "paths": {
...
    "/api/item": {
      "get": {
        "summary": "List all items",
        "operationId": "get_api-item-2490f8ffde47d4c357b7356ded0f88fe",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      },
      "put": {
        "summary": "Create a new item",
        "operationId": "put_api-item-2490f8ffde47d4c357b7356ded0f88fe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Editor": []
          }
        ]
      }
    },
...
    "/api/item/{id}": {
      "get": {
        "summary": "Get an item by id",
        "operationId": "get_api-item-446b04a6c9093bc63aa458a66fe2b2f1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "ItemId",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      },
      "put": {
        "summary": "Create or update an item with id",
        "operationId": "put_api-item-446b04a6c9093bc63aa458a66fe2b2f1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "ItemId",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Editor": []
          }
        ]
      },
      "delete": {
        "summary": "Delete an item",
        "operationId": "delete_api-item-446b04a6c9093bc63aa458a66fe2b2f1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "ItemId",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Editor": []
          }
        ]
      }
    },
...
    "/api/relation": {
      "get": {
        "summary": "List all relations",
        "operationId": "get_api-relation-a6204db2d27ad54d2a69cd7e9d39c398",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      },
      "put": {
        "summary": "Create a relation",
        "operationId": "put_api-relation-a6204db2d27ad54d2a69cd7e9d39c398",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Relation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Editor": []
          }
        ]
      }
    },
    "/api/relation/{id}": {
      "put": {
        "summary": "Create or update a relation by id",
        "operationId": "put_api-relation-c74778c7c61ae2c84457a9e29909fa7f",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "RelationId",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Relation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Editor": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a relation",
        "operationId": "delete_api-relation-c74778c7c61ae2c84457a9e29909fa7f",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "ItemId",
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Editor": []
          }
        ]
      }
    },
...
    "/api/types/item": {
      "get": {
        "summary": "List item types",
        "operationId": "get_api-types-item-c2687ce20abda1bdebbad3f33bdaf745",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "AbsItemTypeIds",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Absolute_ItemTypeId"
                  },
                  "uniqueItems": true
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/item/{package}": {
      "get": {
        "summary": "List item types for a package",
        "operationId": "get_api-types-item-2dab86969b5f7f4684e5f062e0400222",
        "parameters": [
          {
            "name": "package",
            "in": "path",
            "required": true,
            "schema": {
              "title": "PackageId",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ItemTypeIds",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemTypeId"
                  },
                  "uniqueItems": true
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/item/{package}/{type}": {
      "get": {
        "summary": "Get info on an item type",
        "operationId": "get_api-types-item-2578a64cb06d2ce54727fe031ee41f07",
        "parameters": [
          {
            "name": "package",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PackageId"
            }
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ItemTypeId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemTypeInfo"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/property": {
      "get": {
        "summary": "List property types",
        "operationId": "get_api-types-property-ddf577326ba19d392e35bf3b66f458a1",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/property/{package}": {
      "get": {
        "summary": "List property types for a package",
        "operationId": "get_api-types-property-619e5532f65563b7b27d393f566dba7c",
        "parameters": [
          {
            "name": "package",
            "in": "path",
            "required": true,
            "schema": {
              "title": "PackageId",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/property/{package}/{type}": {
      "get": {
        "summary": "Get info on a property type",
        "operationId": "get_api-types-property-88c76de994e539b6f9a653142f7ba9da",
        "parameters": [
          {
            "name": "package",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PackageId"
            }
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PropertyId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyType"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/relation": {
      "get": {
        "summary": "List relation types",
        "operationId": "get_api-types-relation-5b3e28ff8d56b4aa49f571d2cf54f546",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "AbsRelationTypeIds",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Absolute_RelationTypeId"
                  },
                  "uniqueItems": true
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/relation/{package}": {
      "get": {
        "summary": "List item types for a package",
        "operationId": "get_api-types-relation-bb9e24f873ff8a3857e9a50eeaf9c17a",
        "parameters": [
          {
            "name": "package",
            "in": "path",
            "required": true,
            "schema": {
              "title": "PackageId",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "RelationTypeIds",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelationTypeId"
                  },
                  "uniqueItems": true
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/relation/{package}/{type}": {
      "get": {
        "summary": "Get info on an item type",
        "operationId": "get_api-types-relation-cd8cad06576d7e0459b0f2bf765e8c3e",
        "parameters": [
          {
            "name": "package",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PackageId"
            }
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RelationTypeId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationTypeInfo"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
    "/api/types/selector": {
      "post": {
        "summary": "Get info for selector autocompletion",
        "operationId": "post_api-types-selector-951ae1f4ca09e0ea0b9ddab76ff4f80b",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectorCtx"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectorCtxInfo"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "Viewer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Absolute_ItemTypeId": {
        "type": "string"
      },
...
      "Absolute_RelationTypeId": {
        "type": "string"
      },
...
      "Item": {
        "type": "object",
        "properties": {
          "assert": {
            "default": {},
            "allOf": [
              {
                "$ref": "#/components/schemas/MetricSelector"
              }
            ]
          },
          "fork": {
            "default": false,
            "type": "boolean"
          },
          "items": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemRef"
            },
            "uniqueItems": true
          },
          "keys": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabelName"
            },
            "uniqueItems": true
          },
          "metrics": {
            "default": {},
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Metric"
            }
          },
          "query": {
            "default": {},
            "allOf": [
              {
                "$ref": "#/components/schemas/MetricSelector"
              }
            ]
          }
        }
      },
...
      "ItemName": {
        "type": "string"
      },
      "ItemRef": {
        "type": "object",
        "required": [
          "item"
        ],
        "properties": {
          "item": {
            "$ref": "#/components/schemas/ItemName"
          },
          "module": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ModuleName"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "ItemTypeId": {
        "type": "string"
      },
...
      "ItemTypeInfo": {
        "title": "ItemTypeInfo",
        "type": "object",
        "required": [
          "children",
          "implementors",
          "implements",
          "keys",
          "name",
          "parents",
          "prometheus_metrics",
          "properties",
          "source_of",
          "target_of"
        ],
        "properties": {
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relative_ItemTypeId"
            },
            "uniqueItems": true
          },
          "implementors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relative_ItemTypeId"
            },
            "uniqueItems": true
          },
          "implements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relative_ItemTypeId"
            },
            "uniqueItems": true
          },
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relative_PropertyId"
            },
            "uniqueItems": true
          },
          "name": {
            "$ref": "#/components/schemas/ItemTypeName"
          },
          "name_template": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RelativeNameTemplate"
              }
            ],
            "nullable": true
          },
          "parents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relative_ItemTypeId"
            },
            "uniqueItems": true
          },
          "prometheus_metrics": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "$ref": "#/components/schemas/ItemMetrics"
              }
            }
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/PropertyType"
            }
          },
          "source_of": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relative_RelationTypeId"
            },
            "uniqueItems": true
          },
          "target_of": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relative_RelationTypeId"
            },
            "uniqueItems": true
          }
        }
      },
...
      "LabelName": {
        "description": "Prometheus label name. Must match regex /[a-zA-Z_][a-zA-Z0-9_]*/.",
        "type": "string"
      },
...
      "MetricSelector": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/LabelSelector"
        }
      },
...
      "PackageId": {
        "type": "string"
      },
...
      "PropertyId": {
        "type": "string"
      },
...
      "PropertyType": {
        "title": "PropertyType",
        "type": "object",
        "required": [
          "name",
          "value"
        ],
        "properties": {
          "description": {
            "default": null,
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "value": {}
        }
      },
...
      "Relation": {
        "type": "object",
        "required": [
          "relation_type",
          "source",
          "target"
        ],
        "properties": {
          "properties": {
            "default": {},
            "type": "object",
            "additionalProperties": true
          },
          "relation_type": {
            "$ref": "#/components/schemas/Relative_RelationTypeId"
          },
          "source": {
            "$ref": "#/components/schemas/ItemId"
          },
          "target": {
            "$ref": "#/components/schemas/ItemId"
          }
        }
      },
...
      "RelationTypeId": {
        "type": "string"
      },
...
      "RelationTypeInfo": {
        "title": "RelationTypeInfo",
        "type": "object",
        "required": [
          "bidirectional",
          "multiplicity",
          "name",
          "prometheus_metrics",
          "properties",
          "source",
          "target"
        ],
        "properties": {
          "bidirectional": {
            "type": "boolean"
          },
          "connector": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Connector"
              }
            ],
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "multiplicity": {
            "$ref": "#/components/schemas/Multiplicity"
          },
          "name": {
            "type": "string"
          },
          "prometheus_metrics": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "$ref": "#/components/schemas/RelationMetrics"
              }
            }
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/PropertyType"
            }
          },
          "source": {
            "$ref": "#/components/schemas/Relative_ItemTypeId"
          },
          "target": {
            "$ref": "#/components/schemas/Relative_ItemTypeId"
          }
        }
      },
...
      "SelectorCtx": {
        "title": "SelectorCtx",
        "oneOf": [
          {
            "title": "item",
            "type": "object",
            "required": [
              "item"
            ],
            "properties": {
              "item": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Absolute_ItemTypeId"
                },
                "uniqueItems": true,
                "nullable": true
              }
            },
            "additionalProperties": false
          },
          {
            "title": "relation",
            "type": "object",
            "required": [
              "relation"
            ],
            "properties": {
              "relation": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Absolute_RelationTypeId"
                },
                "uniqueItems": true,
                "nullable": true
              }
            },
            "additionalProperties": false
          }
        ]
      },
      "SelectorCtxInfo": {
        "title": "SelectorCtxInfo",
        "type": "object",
        "required": [
          "items",
          "parents",
          "properties",
          "relations",
          "sources",
          "targets"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Absolute_ItemTypeId"
            },
            "uniqueItems": true
          },
          "parents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Absolute_ItemTypeId"
            },
            "uniqueItems": true
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Absolute_PropertyId"
            },
            "uniqueItems": true
          },
          "relations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Absolute_RelationTypeId"
            },
            "uniqueItems": true
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Absolute_ItemTypeId"
            },
            "uniqueItems": true
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Absolute_ItemTypeId"
            },
            "uniqueItems": true
          }
        }
      },
...
    }
  }
}